Sorry, this is a crappy commit, lots of stuff. Whitespace cleanup.
Major change:
Refactored how routes are loaded.  There is now a routes directory that contains a routeLoader which loads in all of the Collections from files in "routes/Collections".  TODO: Readme needs an update.

Major change:
Updated Apache puppet module to allow for Allow Override: all within the puppet manifest.  This fixes an issue with the defaul Apache settings caused in the upgrade of Ubuntu 12.04 from 2.3.x to 2.4.
Additionally, changed php ppa to php5-oldstable, which anchors php to 5.4 instead of 5.5, since 5.5 is not fully compatable with my puppet config.